chatbot-prod/pyproject.toml
[tool.poetry]
name = "chatbot-prod"
version = "0.1.0"
description = ""
authors = ["Samuel Yawson <s.yawson@4th-ir.com>"]
readme = "README.md"
packages = [{include = "chatbot_prod"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
pydantic = "^1.10.7"
python-jose = "^3.3.0"
pinecone-client = "^2.2.1"
langchain = "^0.0.176"
motor = "^3.1.2"
pymongo = "^4.3.3"
transformers = "^4.29.2"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
openai = "^0.27.7"
bcrypt = "^4.0.1"
pymupdf = "^1.22.3"
tiktoken = "^0.4.0"
torch = "^2.0.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
